+++ Tasks/Tasks/nlog.config
@@ -10,7 +10,7 @@
 	<targets>
 		<target xsi:type="Database"
 		name="DB"
-		connectionString="Server=DESKTOP-B61M7D6;Database=NefeshYehudi;Trusted_Connection=True;Trust Server Certificate=true;"
+		connectionString="Server=srv2\\pupils;Database=NefeshYehudi;Trusted_Connection=True;Trust Server Certificate=true;"
 		commandText="INSERT INTO Logs (CreatedOn,Message,Level,StackTrace,Exception,Logger,Url) VALUES (@datetime,@msg,@level,@trace,@exception,@logger,@url)"
 			keepConnection="true">
 			<parameter name="@datetime" layout="${date}" />
