--- ny-tasks/src/app/tasks/tasks.component.html
@@ -1,1 +1,0 @@
-<p>tasks works!</p>
